+++ tests/ethernet/EthMacCore/test_EthMacTxExport.py
@@ -9,26 +9,26 @@
 ##############################################################################
 
 # Test methodology:
-# - Sweep: Cover both supported functional PHY mappings in this wrapper,
-#   `GMII` and `XGMII`, and include a link-not-ready interval in each run.
+# - Sweep: Cover both supported functional PHY mappings, `GMII` and `XGMII`,
+#   and explicitly exercise the exporter with `phyReady` both high and low.
 # - Stimulus: Send one minimum-size frame, one longer multi-beat frame, one
-#   frame while `phyReady=0`, and then one recovery frame after re-enabling the
-#   link.
-# - Checks: Ready PHY modes must recover the original AXIS bytes and pulse
-#   `rxCountEn`, the blocked frame must not appear while the link is down, and
-#   the receiver must return to normal operation after `phyReady` is restored.
-#   The recovery expectation is PHY-specific: GMII drops traffic presented
-#   while the link is down, while the current XGMII loopback path presents that
-#   queued frame once `phyReady` returns.
-# - Timing: GMII takes many more cycles than XGMII to serialize a frame, so the
-#   bench scales its receive timeout to the selected PHY mode.
+#   frame while the link is marked not ready, and then one recovery frame after
+#   the link returns.
+# - Checks: Successful transmissions must pulse `txCountEn`, the blocked frame
+#   must raise `txLinkNotReady` without producing output data while the link is
+#   down, and normal export behavior must recover cleanly after the ready
+#   signal is restored. The recovery expectation is PHY-specific because the
+#   current XGMII path drains the blocked frame after the link returns, while
+#   the GMII path drops it.
+# - Timing: The receive timeout is scaled to the chosen PHY because the GMII
+#   path serializes one byte per clock while XGMII transmits eight.
 
 import cocotb
 import os
 import pytest
 
 from tests.common.regression_utils import parameter_case, run_surf_vhdl_test
-from tests.ethernet.eth_mac.ethmac_test_utils import (
+from tests.ethernet.EthMacCore.ethmac_test_utils import (
     ETHMAC_RTL_SOURCES,
     build_ethernet_frame,
     cycle,
@@ -47,7 +47,7 @@ WRAPPER_PATH = "ethernet/EthMacCore/wrappers/EthMacImportExportLoopbackWrapper.v
 
 
 @cocotb.test()
-async def eth_mac_rx_import_test(dut):
+async def eth_mac_tx_export_test(dut):
     phy_type = os.environ["PHY_TYPE_G"]
     timeout_cycles = 512 if phy_type == "XGMII" else 4096
 
@@ -66,30 +66,30 @@ async def eth_mac_rx_import_test(dut):
     assert sink is not None
 
     min_frame = build_ethernet_frame(
-        dst_mac=0x020304050607,
-        src_mac=0x0A0B0C0D0E0F,
-        eth_type=0x88B5,
+        dst_mac=0x0C0D0E0F1011,
+        src_mac=0x121314151617,
+        eth_type=0x9000,
         payload=bytes(range(46)),
     )
-    min_pulse = cocotb.start_soon(wait_signal_pulse(dut.rxCountEn, clk=bench.clk, timeout_cycles=timeout_cycles))
+    min_pulse = cocotb.start_soon(wait_signal_pulse(dut.txCountEn, clk=bench.clk, timeout_cycles=timeout_cycles))
     min_send = cocotb.start_soon(
         send_contiguous_frame(source, frame_beats_from_bytes(min_frame), clk=bench.clk)
     )
     min_observed = await recv_frame(sink, clk=bench.clk, timeout_cycles=timeout_cycles)
     await min_send
     await min_pulse
-    assert payload_from_beats(min_observed) == min_frame
-    assert int(dut.rxCrcError.value) == 0
 
-    # Drive a second, longer frame so the import path is checked across a
-    # multi-beat packet instead of only a minimum-sized transfer.
+    assert payload_from_beats(min_observed) == min_frame
+    assert int(dut.txUnderRun.value) == 0
+    assert int(dut.txLinkNotReady.value) == 0
+
     long_frame = build_ethernet_frame(
-        dst_mac=0x111213141516,
-        src_mac=0x212223242526,
+        dst_mac=0x202122232425,
+        src_mac=0x303132333435,
         eth_type=0x88B5,
         payload=bytes(range(96)),
     )
-    long_pulse = cocotb.start_soon(wait_signal_pulse(dut.rxCountEn, clk=bench.clk, timeout_cycles=timeout_cycles))
+    long_pulse = cocotb.start_soon(wait_signal_pulse(dut.txCountEn, clk=bench.clk, timeout_cycles=timeout_cycles))
     long_send = cocotb.start_soon(
         send_contiguous_frame(source, frame_beats_from_bytes(long_frame), clk=bench.clk)
     )
@@ -97,35 +97,37 @@ async def eth_mac_rx_import_test(dut):
     await long_send
     await long_pulse
     assert payload_from_beats(long_observed) == long_frame
-    assert int(dut.rxCrcError.value) == 0
+    assert int(dut.txUnderRun.value) == 0
 
-    # A deasserted PHY-ready input resets the import logic, so traffic that
-    # arrives in that interval must not leak partial output while the link is
-    # down. The current XGMII path replays the blocked frame after recovery,
-    # whereas the GMII path discards it.
+    # When the PHY is not ready, the exporter must flag the condition and avoid
+    # emitting partial output while the link is down. The exact post-recovery
+    # behavior is PHY-specific and is checked below.
     dut.phyReady.value = 0
     blocked_frame = build_ethernet_frame(
-        dst_mac=0x313233343536,
-        src_mac=0x414243444546,
+        dst_mac=0x404142434445,
+        src_mac=0x505152535455,
         eth_type=0x9000,
         payload=bytes(range(32)),
     )
+    blocked_pulse = cocotb.start_soon(wait_signal_pulse(dut.txLinkNotReady, clk=bench.clk, timeout_cycles=timeout_cycles))
     blocked_send = cocotb.start_soon(
         send_contiguous_frame(source, frame_beats_from_bytes(blocked_frame), clk=bench.clk)
     )
     await blocked_send
+    await blocked_pulse
     await expect_no_output(sink, clk=bench.clk, cycles=(24 if phy_type == "XGMII" else 128))
+    assert int(dut.txUnderRun.value) == 0
 
     dut.phyReady.value = 1
     await cycle(bench.clk, 4)
 
     recovery_frame = build_ethernet_frame(
-        dst_mac=0x515253545556,
-        src_mac=0x616263646566,
+        dst_mac=0x606162636465,
+        src_mac=0x707172737475,
         eth_type=0x88B5,
-        payload=b"link-recovery-frame" + bytes(27),
+        payload=b"post-link-recovery" + bytes(28),
     )
-    recovery_pulse = cocotb.start_soon(wait_signal_pulse(dut.rxCountEn, clk=bench.clk, timeout_cycles=timeout_cycles))
+    recovery_pulse = cocotb.start_soon(wait_signal_pulse(dut.txCountEn, clk=bench.clk, timeout_cycles=timeout_cycles))
     recovery_send = cocotb.start_soon(
         send_contiguous_frame(source, frame_beats_from_bytes(recovery_frame), clk=bench.clk)
     )
@@ -134,19 +136,18 @@ async def eth_mac_rx_import_test(dut):
     await recovery_pulse
 
     if phy_type == "XGMII":
-        # The XGMII export/import loopback retains the blocked frame across the
-        # ready transition, so the first recovered packet is the stalled one.
-        # Because that packet traverses the TX path, it comes back padded to
-        # Ethernet's minimum 60-byte frame size.
+        # In the XGMII path the held-off frame drains first once the link
+        # returns, so the recovery frame arrives second. The stalled frame is
+        # padded by the TX path up to Ethernet's minimum non-FCS size.
         assert payload_from_beats(recovery_observed) == pad_ethernet_frame_to_min_size(blocked_frame)
         drained_observed = await recv_frame(sink, clk=bench.clk, timeout_cycles=timeout_cycles)
         assert payload_from_beats(drained_observed) == recovery_frame
     else:
-        # The GMII path drops the blocked frame entirely, so the next output is
-        # the fresh recovery packet.
+        # GMII drops the frame that arrived while the link was down.
         assert payload_from_beats(recovery_observed) == recovery_frame
 
-    assert int(dut.rxCrcError.value) == 0
+    assert int(dut.txUnderRun.value) == 0
+    assert int(dut.txLinkNotReady.value) == 0
 
 
 PARAMETER_SWEEP = [
@@ -156,7 +157,7 @@ PARAMETER_SWEEP = [
 
 
 @pytest.mark.parametrize("parameters", PARAMETER_SWEEP)
-def test_EthMacRxImport(parameters):
+def test_EthMacTxExport(parameters):
     run_surf_vhdl_test(
         test_file=__file__,
         toplevel="surf.ethmacimportexportloopbackwrapper",
